タグ4bit化(PR #75/#77/#78)を取り込む。競合は Makefile と
test/c/runtime_test.c の2ファイル。

- Makefile: 純粋な union。compiler-optimization 側の disasm.c/disasm_lisp.c
  (SRC)と disasm.h/disasm_lisp.h(HDR)、main 側の za_jit_tags.h(HDR)は
  互いに別物なので両方残す。za_jit_tags.h が落ちると JIT のタグ定数が
  TAG_MASK から追随しなくなるため、4bit化が有効であるために必須
- runtime_test.c 後半: 互いに独立した新規テスト関数の union。
  compiler-optimization 側の disassembler アドレス分類テスト6本と、
  main 側の16byte境界/タグのテスト3本。main() の登録は自動マージ済み
- runtime_test.c の SMALL_HEAP_SIZE: 下記

SMALL_HEAP_SIZE は 82KB にした。両ブランチが別々の理由で別々の値にしており、
**どちらの値もマージ後にはハングする**(実際に80KBで runtime_test が
os_panic の無限ループに入りCPU100%のまま止まらなくなった)。
declaim(Phase2)の追加分と16byte切り上げの+416byteが足し合わさり、
窓ごと1KB上へずれたためである。

マージ後のツリーで掃引した実測値:
  80KB: ハング / 81〜84KB: 5395 OK 0 NG / 85KB以降: GC未発火でNG
窓 81〜84KB の中央を採って 82KB。両端で壊れ方が違い(下はハング、上はNG)、
ハングのほうが切り分けに時間がかかるので下側には寄せない。

検証:
- make test: 8403 OK / 0 NG
- make build(実機PE32+): 成功

+++ src/c/reader.c
@@ -649,7 +649,8 @@ static lisp_val_t read_char_literal(reader_source_t *src) {
     }
 
     if (len == 1) {
-        return os_make_char(buf[0]);
+        /* [4bit化] bufはsigned char配列。UINT8を経由して符号拡張を避ける */
+        return os_make_char((UINT8)buf[0]);
     }
     if (char_name_matches(buf, len, "SPACE")) {
         return os_make_char(' ');
